# backend/src/schemas/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Electro Atlas API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://electro-atlas-backend-production.up.railway.app/api/v1",
      "description": "Production server"
    }
  ],
...
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Register a new user",
        "operationId": "registerUser",
        "tags": [
          "auth"
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/RegisterUser"
        },
        "responses": {
          "204": {
            "description": "User registered successfully"
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "409": {
            "$ref": "#/components/responses/ConflictError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login a user",
        "operationId": "loginUser",
        "tags": [
          "auth"
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/UserLogin"
        },
        "responses": {
          "200": {
            "description": "User logged in successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserLoginResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/auth/refresh-token": {
      "post": {
        "summary": "Refresh access token",
        "operationId": "refreshToken",
        "tags": [
          "auth"
        ],
        "security": [
          {
            "RefreshTokenCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Token refreshed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshTokenResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/auth/verify-email/confirm": {
      "post": {
        "summary": "Confirm email verification",
        "operationId": "verifyEmail",
        "tags": [
          "auth"
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/VerifyEmail"
        },
        "responses": {
          "204": {
            "description": "Email verified successfully"
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/auth/verify-email/request": {
      "post": {
        "summary": "Request email verification code",
        "operationId": "resendVerificationCode",
        "tags": [
          "auth"
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/ResendVerificationCode"
        },
        "responses": {
          "204": {
            "description": "Verification code resent successfully"
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/auth/password-reset/request": {
      "post": {
        "summary": "Request password reset",
        "operationId": "forgotPassword",
        "tags": [
          "auth"
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/ForgotPassword"
        },
        "responses": {
          "200": {
            "description": "Password reset link sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForgotPasswordResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/auth/password-reset/confirm": {
      "post": {
        "summary": "Confirm reset password",
        "operationId": "resetPassword",
        "tags": [
          "auth"
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/ResetPassword"
        },
        "responses": {
          "204": {
            "description": "Password reset successfully"
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/auth/oauth/google": {
      "get": {
        "summary": "Initiate Google Authentication",
        "operationId": "initiateGoogleAuth",
        "tags": [
          "auth"
        ],
        "parameters": [
          {
            "name": "callbackUrl",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "URL to redirect after successful authentication"
          }
        ],
        "responses": {
          "302": {
            "description": "URL of Google's authorization endpoint"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "requestBodies": {
      "RegisterUser": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "email",
                "password",
                "confirmPassword"
              ],
              "properties": {
                "email": {
                  "type": "string",
                  "format": "email"
                },
                "password": {
                  "type": "string",
                  "minLength": 8
                },
                "confirmPassword": {
                  "type": "string"
                },
                "wishlist": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "productId"
                    ],
                    "properties": {
                      "productId": {
                        "type": "string"
                      }
                    }
                  }
                },
                "cartItems": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "quantity",
                      "productId"
                    ],
                    "properties": {
                      "quantity": {
                        "type": "number",
                        "minimum": 1
                      },
                      "productId": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "UserLogin": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "email",
                "password"
              ],
              "properties": {
                "email": {
                  "type": "string",
                  "format": "email"
                },
                "password": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "VerifyEmail": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "email",
                "verificationCode"
              ],
              "properties": {
                "email": {
                  "type": "string",
                  "format": "email"
                },
                "verificationCode": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "ResendVerificationCode": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "email"
              ],
              "properties": {
                "email": {
                  "type": "string",
                  "format": "email"
                }
              }
            }
          }
        }
      },
      "ForgotPassword": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "email"
              ],
              "properties": {
                "email": {
                  "type": "string",
                  "format": "email"
                }
              }
            }
          }
        }
      },
      "ResetPassword": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "resetToken",
                "password",
                "confirmPassword"
              ],
              "properties": {
                "resetToken": {
                  "type": "string"
                },
                "password": {
                  "type": "string",
                  "minLength": 8
                },
                "confirmPassword": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
    },
    "schemas": {
      "UserLoginResponse": {
        "type": "object",
        "required": [
          "accessToken"
        ],
        "properties": {
          "accessToken": {
            "type": "string"
          }
        }
      },
      "RefreshTokenResponse": {
        "type": "object",
        "required": [
          "accessToken"
        ],
        "properties": {
          "accessToken": {
            "type": "string"
          }
        }
      },
      "ForgotPasswordResponse": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "ErrorModel": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "errors": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "AccessTokenAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token must be sent in the Authorization header in the format:\nBearer <token>\n"
      },
      "RefreshTokenCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jid",
        "description": "Refresh token must be sent in the 'jid' cookie"
      }
    },
...
    "responses": {
      "InternalServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorModel"
            }
          }
        }
      },
      "NotFoundError": {
        "description": "Not found error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorModel"
            }
          }
        }
      },
      "BadRequestError": {
        "description": "Bad request error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorModel"
            }
          }
        }
      },
      "UnauthorizedError": {
        "description": "Unauthorized error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorModel"
            }
          }
        }
      },
      "ForbiddenError": {
        "description": "Forbidden error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorModel"
            }
          }
        }
      },
      "ConflictError": {
        "description": "Conflict error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorModel"
            }
          }
        }
      }
    }
  }
}